Added support for negative enum values

+++ src/Symbols.h
@@ -85,7 +85,7 @@ struct Symbol
 		bool isUnion = false;
 	} pack;
 
-	int enumValue;
+	int64_t enumValue;
 
 	TokenList macroTokens;
